rename Image2 to Image and bump sdk package version
