+++ packages/nouns-sdk/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@nouns/sdk",
-  "version": "0.1.6",
+  "version": "0.1.7",
   "description": "Useful tooling for interacting with the Nouns protocol",
   "author": "Nounders",
   "homepage": "https://nouns.wtf",
